feat: validate if laboratory exists before delete

+++ src/services/LaboratoriesServices/index.ts
@@ -1,6 +1,7 @@
 import { CreateLaboratoryDTO } from '@src/dtos/LaboratoriesDTOS';
 import { Laboratory, LaboratoryStatus } from '@src/entities/Laboratory';
 import { LaboratoriesRepositories } from '@src/repositories/LaboratoriesRepositories';
+import { NotFoundError } from '@src/utils/errors/NotFoundError';
 import { MakePartial } from '@src/utils/generics/MakePartial';
 import { logger } from '@src/utils/logger';
 
@@ -26,7 +27,20 @@ export class LaboratoriesServices {
     return await this.laboratoriesRepositories.findAll();
   }
 
+  async findById(id: string): Promise<Laboratory | null> {
+    const laboratory = await this.laboratoriesRepositories.findById(id);
+
+    return laboratory;
+  }
+
   async deleteById(id: string): Promise<Laboratory> {
-    return await this.laboratoriesRepositories.deleteById(id);
+    const thisLaboratoryExists = await this.findById(id);
+
+    if (!thisLaboratoryExists) {
+      throw new NotFoundError('Laboratory not found.');
+    }
+
+    const removedLaboratoy = await this.laboratoriesRepositories.deleteById(id);
+    return removedLaboratoy;
   }
 }
